backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.1
pypdf==5.2.0
nltk==3.9.1
python-dotenv==1.0.1
gunicorn==23.0.0
scikit-learn==1.6.1
numpy==2.2.3
scipy==1.15.2
google-generativeai==0.8.6
pydantic==2.10.6
pytest==8.3.4
pytest-flask==1.3.0